Bumped regions from SectorModelReader to SosModelReader
