+++ smif/data_layer/sector_model_config.py
@@ -3,9 +3,6 @@
 """
 import logging
 import os
-from glob import glob
-
-import fiona
 
 from .load import load
 from .validate import validate_interventions
@@ -63,9 +60,6 @@ class SectorModelReader(object):
 
         self.inputs = None
         self.outputs = None
-        self.time_intervals = None
-        self.regions = None
-
         self.initial_conditions = None
         self.interventions = None
 
@@ -74,8 +68,6 @@ class SectorModelReader(object):
         """
         self.inputs = self.load_inputs()
         self.outputs = self.load_outputs()
-
-        self.regions = self.load_regions()
         self.initial_conditions = self.load_initial_conditions()
         self.interventions = self.load_interventions()
 
@@ -119,8 +111,6 @@ class SectorModelReader(object):
             "classname": self.model_classname,
             "inputs": self.inputs,
             "outputs": self.outputs,
-            "time_intervals": self.time_intervals,
-            "regions": self.regions,
             "initial_conditions": self.initial_conditions,
             "interventions": self.interventions
         }
@@ -175,26 +165,4 @@ class SectorModelReader(object):
             new_data = load(path)
             validate_interventions(new_data, path)
             data.extend(new_data)
-        return data
-
-    def load_regions(self):
-        """Model regions are specified in ``data/<sectormodel>/regions.*``
-
-        The file format must be possible to parse with GDAL, and must contain
-        an attribute "name" to use as an identifier for the region.
-        """
-        path = os.path.join(self.model_config_dir, 'regions.shp')
-
-        if not os.path.exists(path):
-            paths = glob("{}/regions.*".format(self.model_config_dir))
-            if len(paths) == 1:
-                path = paths[0]
-            else:
-                msg = "regions config file not found for {} model"
-                raise FileNotFoundError(msg.format(self.model_name))
-
-        with fiona.drivers():
-            with fiona.open(path) as src:
-                data = [f for f in src]
-
         return data
